Generative-AI/Youtube_Rag_Chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
youtube-transcript-api==1.1.1
langchain
langchain-community
langchain-core
langchain-google-genai
sentence-transformers
faiss-cpu
streamlit